Move useEditCalendar to TimeTableStep
